requirements.txt
altgraph==0.17
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-crispy-forms==1.9.0
django-filter==2.2.0
django-registration-redux==2.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.3
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
patsy==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
pygame==1.9.6
Pygments==2.5.2
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.6
pypiwin32==223
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
statsmodels==0.11.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
